# fix: faster diff avoiding twice instantiation of semver objects